pyproject.toml
[project]
name = "llm-sycophancy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "flash-attn",
    "python-dotenv>=1.2.1",
    "setuptools",
    "torch>=2.10.0",
    "transformers>=5.1.0",
    "vllm",
]
